refactor: move ellipsify helper to ui package
