+++ packages/portfolio/src/ui/portfolio-ui-tokens.tsx
@@ -1,8 +1,8 @@
 import type { Cluster } from '@workspace/db/entity/cluster'
 import type { GetTokenAccountsResult } from '@workspace/solana-client/get-token-accounts'
 
-import { ellipsify } from '@workspace/core/ellipsify'
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@workspace/ui/components/table'
+import { ellipsify } from '@workspace/ui/lib/ellipsify'
 
 import { PortfolioUiExplorerLink } from './portfolio-ui-explorer-link.js'
