+++ packages/settings/src/data-access/use-derive-and-create-wallet.tsx
@@ -2,6 +2,7 @@ import type { Account } from '@workspace/db/entity/account'
 
 import { useMutation } from '@tanstack/react-query'
 import { useDbWalletCreate } from '@workspace/db-react/use-db-wallet-create'
+import { ellipsify } from '@workspace/ui/lib/ellipsify'
 
 import { useDeriveFromMnemonic } from './use-derive-from-mnemonic.js'
 
@@ -21,6 +22,7 @@ export function useDeriveAndCreateWallet() {
           ...derivedWallet,
           accountId: item.id,
           derivationIndex: index,
+          name: ellipsify(derivedWallet.publicKey),
           type: 'Derived',
         },
       })
