+++ packages/ui/src/lib/ellipsify.ts
